Move ServerAppComponents
